Remove respository
