+++ sdk/src/main/java/com/microsoft/did/sdk/IssuanceManager.kt
@@ -6,22 +6,32 @@ import com.microsoft.did.sdk.credential.models.VerifiableCredential
 import com.microsoft.did.sdk.credential.service.IssuanceRequest
 import com.microsoft.did.sdk.credential.service.IssuanceResponse
 import com.microsoft.did.sdk.credential.service.RequestedVcMap
+import com.microsoft.did.sdk.credential.service.protectors.IssuanceResponseFormatter
+import com.microsoft.did.sdk.datasource.network.apis.ApiProvider
+import com.microsoft.did.sdk.datasource.network.credentialOperations.FetchContractNetworkOperation
+import com.microsoft.did.sdk.datasource.network.credentialOperations.SendVerifiableCredentialIssuanceRequestNetworkOperation
 import com.microsoft.did.sdk.datasource.repository.VerifiableCredentialRepository
 import com.microsoft.did.sdk.identifier.models.Identifier
+import com.microsoft.did.sdk.util.Constants
 import com.microsoft.did.sdk.util.controlflow.Result
 import com.microsoft.did.sdk.util.controlflow.runResultTry
+import com.microsoft.did.sdk.util.formVerifiableCredential
+import com.microsoft.did.sdk.util.serializer.Serializer
 import javax.inject.Inject
 import javax.inject.Singleton
 
 @Singleton
 class IssuanceManager @Inject constructor(
     private val identifierManager: IdentifierManager,
-    private val vcRepository: VerifiableCredentialRepository
+    private val exchangeService: ExchangeService,
+    private val apiProvider: ApiProvider,
+    private val issuanceResponseFormatter: IssuanceResponseFormatter,
+    private val serializer: Serializer
 ) {
 
     suspend fun getIssuanceRequest(contractUrl: String): Result<IssuanceRequest> {
         return runResultTry {
-            val contract = vcRepository.getContract(contractUrl).abortOnError()
+            val contract = getContract(contractUrl).abortOnError()
             val request = IssuanceRequest(contract, contractUrl)
             Result.Success(request)
         }
@@ -37,10 +47,10 @@ class IssuanceManager @Inject constructor(
                 val pairwiseIdentifier =
                     identifierManager.createPairwiseIdentifier(masterIdentifier, response.request.entityIdentifier).abortOnError()
                 val requestedVcMap = exchangeVcsInIssuanceRequest(response, pairwiseIdentifier).abortOnError()
-                vcRepository.sendIssuanceResponse(response, pairwiseIdentifier, requestedVcMap).abortOnError()
+                sendIssuanceResponse(response, pairwiseIdentifier, requestedVcMap).abortOnError()
             } else {
                 val requestedVcMap = response.requestedVcMap
-                vcRepository.sendIssuanceResponse(response, masterIdentifier, requestedVcMap).abortOnError()
+                sendIssuanceResponse(response, masterIdentifier, requestedVcMap).abortOnError()
             }
             Result.Success(verifiableCredential)
         }
@@ -53,9 +63,38 @@ class IssuanceManager @Inject constructor(
         return runResultTry {
             val exchangedVcMap = response.requestedVcMap.mapValues {
                 val owner = identifierManager.getIdentifierById(it.value.contents.sub).abortOnError()
-                vcRepository.getExchangedVerifiableCredential(it.value, owner, pairwiseIdentifier).abortOnError()
+                exchangeService.getExchangedVerifiableCredential(it.value, owner, pairwiseIdentifier).abortOnError()
             }
             Result.Success(exchangedVcMap as RequestedVcMap)
+        }
+    }
+
+    private suspend fun getContract(url: String) = FetchContractNetworkOperation(
+        url,
+        apiProvider
+    ).fire()
+
+    private suspend fun sendIssuanceResponse(
+        response: IssuanceResponse,
+        responder: Identifier,
+        requestedVcMap: RequestedVcMap,
+        expiryInSeconds: Int = Constants.DEFAULT_EXPIRATION_IN_SECONDS
+    ): Result<VerifiableCredential> {
+        val formattedResponse = issuanceResponseFormatter.formatResponse(
+            requestedVcMap = requestedVcMap,
+            issuanceResponse = response,
+            responder = responder,
+            expiryInSeconds = expiryInSeconds
+        )
+        val rawVerifiableCredentialResult = SendVerifiableCredentialIssuanceRequestNetworkOperation(
+            response.audience,
+            formattedResponse,
+            apiProvider
+        ).fire()
+
+        return when (rawVerifiableCredentialResult) {
+            is Result.Success -> Result.Success(formVerifiableCredential(rawVerifiableCredentialResult.payload, null, serializer))
+            is Result.Failure -> rawVerifiableCredentialResult
         }
     }
 }
