+++ sdk/src/main/java/com/microsoft/did/sdk/datasource/repository/VerifiableCredentialRepository.kt
@@ -43,116 +43,10 @@ class VerifiableCredentialRepository @Inject constructor(
     private val serializer: Serializer
 ) {
     // Card Issuance Methods.
-    suspend fun getContract(url: String) = FetchContractNetworkOperation(
-        url,
-        apiProvider
-    ).fire()
 
-    suspend fun sendIssuanceResponse(
-        response: IssuanceResponse,
-        responder: Identifier,
-        requestedVcMap: RequestedVcMap,
-        expiryInSeconds: Int = DEFAULT_EXPIRATION_IN_SECONDS
-    ): Result<VerifiableCredential> {
-        val formattedResponse = issuanceResponseFormatter.formatResponse(
-            requestedVcMap = requestedVcMap,
-            issuanceResponse = response,
-            responder = responder,
-            expiryInSeconds = expiryInSeconds
-        )
-        val rawVerifiableCredentialResult = SendVerifiableCredentialIssuanceRequestNetworkOperation(
-            response.audience,
-            formattedResponse,
-            apiProvider
-        ).fire()
-
-        return when (rawVerifiableCredentialResult) {
-            is Result.Success -> Result.Success(formVerifiableCredential(rawVerifiableCredentialResult.payload))
-            is Result.Failure -> rawVerifiableCredentialResult
-        }
-    }
 
     // Presentation Methods.
-    suspend fun getRequest(url: String) = FetchPresentationRequestNetworkOperation(url, apiProvider).fire()
 
-    suspend fun sendPresentationResponse(
-        response: PresentationResponse,
-        responder: Identifier,
-        requestedVcPresentationSubmissionMap: RequestedVcPresentationSubmissionMap,
-        expiryInSeconds: Int = DEFAULT_EXPIRATION_IN_SECONDS
-    ): Result<Unit> {
-        val formattedResponse = presentationResponseFormatter.formatResponse(
-            requestedVcPresentationSubmissionMap = requestedVcPresentationSubmissionMap,
-            presentationResponse = response,
-            responder = responder,
-            expiryInSeconds = expiryInSeconds
-        )
-        return SendPresentationResponseNetworkOperation(
-            response.audience,
-            formattedResponse,
-            response.request.content.state,
-            apiProvider
-        ).fire()
-    }
 
-    suspend fun revokeVerifiablePresentation(
-        verifiableCredential: VerifiableCredential,
-        owner: Identifier,
-        rpList: List<String>,
-        reason: String
-    ): Result<RevocationReceipt> {
-        val revocationRequest = RevocationRequest(verifiableCredential, owner, rpList, reason)
-        val formattedRevocationRequest = revocationResponseFormatter.formatResponse(revocationRequest, DEFAULT_EXPIRATION_IN_SECONDS)
-        return sendRevocationRequest(revocationRequest, formattedRevocationRequest)
-    }
 
-    private suspend fun sendRevocationRequest(revocationRequest: RevocationRequest, formattedRevocationRequest: String): Result<RevocationReceipt> {
-        val revocationResult = SendVerifiablePresentationRevocationRequestNetworkOperation(
-            revocationRequest.audience,
-            formattedRevocationRequest,
-            apiProvider,
-            serializer
-        ).fire()
-        return when (revocationResult) {
-            is Result.Success -> revocationResult
-            is Result.Failure -> Result.Failure(RevocationException("Unable to revoke VP"))
-        }
-    }
-
-    suspend fun getExchangedVerifiableCredential(
-        verifiableCredential: VerifiableCredential,
-        owner: Identifier,
-        pairwiseIdentifier: Identifier
-    ): Result<VerifiableCredential> {
-        return sendExchangeRequest(ExchangeRequest(verifiableCredential, pairwiseIdentifier.id, owner), DEFAULT_EXPIRATION_IN_SECONDS)
-    }
-
-    private suspend fun sendExchangeRequest(request: ExchangeRequest, expiryInSeconds: Int): Result<VerifiableCredential> {
-        if (request.audience == "") {
-            throw ExchangeException("Audience is an empty string.")
-        }
-        val formattedPairwiseRequest = exchangeResponseFormatter.formatResponse(request, expiryInSeconds)
-
-        val result = SendVerifiableCredentialIssuanceRequestNetworkOperation(
-            request.audience,
-            formattedPairwiseRequest,
-            apiProvider
-        ).fire()
-
-        return when (result) {
-            is Result.Success -> {
-                val verifiableCredential = formVerifiableCredential(
-                    result.payload,
-                    request.verifiableCredential.picId
-                )
-                Result.Success(verifiableCredential)
-            }
-            is Result.Failure -> result
-        }
-    }
-
-    private fun formVerifiableCredential(rawToken: String, vcId: String? = null): VerifiableCredential {
-        val vcContents = unwrapSignedVerifiableCredential(rawToken, serializer)
-        return VerifiableCredential(vcContents.jti, rawToken, vcContents, vcId ?: vcContents.jti)
-    }
 }
